+++ src/Service/S3/src/Input/CopyObjectRequest.php
@@ -1167,13 +1167,13 @@ final class CopyObjectRequest extends Input
             $headers['x-amz-copy-source-if-match'] = $this->copySourceIfMatch;
         }
         if (null !== $this->copySourceIfModifiedSince) {
-            $headers['x-amz-copy-source-if-modified-since'] = $this->copySourceIfModifiedSince->setTimezone(new \DateTimeZone('GMT'))->format(\DateTimeInterface::RFC7231);
+            $headers['x-amz-copy-source-if-modified-since'] = $this->copySourceIfModifiedSince->setTimezone(new \DateTimeZone('GMT'))->format('D, d M Y H:i:s \G\M\T');
         }
         if (null !== $this->copySourceIfNoneMatch) {
             $headers['x-amz-copy-source-if-none-match'] = $this->copySourceIfNoneMatch;
         }
         if (null !== $this->copySourceIfUnmodifiedSince) {
-            $headers['x-amz-copy-source-if-unmodified-since'] = $this->copySourceIfUnmodifiedSince->setTimezone(new \DateTimeZone('GMT'))->format(\DateTimeInterface::RFC7231);
+            $headers['x-amz-copy-source-if-unmodified-since'] = $this->copySourceIfUnmodifiedSince->setTimezone(new \DateTimeZone('GMT'))->format('D, d M Y H:i:s \G\M\T');
         }
         if (null !== $this->expires) {
             $headers['Expires'] = $this->expires;